# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Coop Salary Board API",
    "description": "API for collection coop salaries anonymously",
    "termsOfService": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:{port}/{basePath}",
      "description": "The Development API server",
      "variables": {
        "port": {
          "default": "3000"
        },
        "basePath": {
          "default": "v1"
        }
      }
    }
  ],
  "components": {
    "schemas": {
...
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "oauth2",
        "description": "Using OAuth2 for password validation flow of generating jwt",
        "flows": {
          "implicit": {
            "authorizationUrl": "/api/v1/login/:jwt"
          }
        }
      }
    },
...
  },
  "paths": {
    "/auth/login": {
      "post": {
        "deprecated": false,
        "summary": "Request a magic link for authentication",
        "security": [],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example1": {
                    "summary": "Example success response",
                    "value": {
                      "message": "Magic link sent to your email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example2": {
                    "summary": "Example error response",
                    "value": {
                      "message": "Email is required"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Database or email error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example3": {
                    "summary": "Example error response",
                    "value": {
                      "message": "Failed to send email"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [],
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "examples": {
                "example1": {
                  "summary": "Example request",
                  "value": "{\"email\": \"user@example.com\"}"
                }
              }
            }
          },
          "description": "User email",
          "required": true
        }
      }
    },
    "/auth/login/{token}": {
      "get": {
        "deprecated": false,
        "summary": "Verify magic link token and authenticate user",
        "security": [],
        "responses": {
          "200": {
            "description": "Success response with access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example1": {
                    "summary": "Example success response",
                    "value": {
                      "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "token_type": "bearer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example2": {
                    "summary": "Example invalid token",
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Magic link token",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/autocomplete/company": {
      "get": {
        "deprecated": false,
        "summary": "Search for companies by name with fuzzy matching",
        "security": [],
        "responses": {
          "200": {
            "description": "List of matching company names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "Example success response",
                    "value": [
                      "Apple Inc",
                      "Apple Technologies",
                      "Appleseed Solutions"
                    ]
                  }
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example2": {
                    "summary": "Example error response",
                    "value": {
                      "message": "Database query failed"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "comp",
            "in": "query",
            "description": "Search parameters for company name",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Search"
        ]
      }
    },
    "/autocomplete/position": {
      "get": {
        "deprecated": false,
        "summary": "Search for positions within a specific company",
        "security": [],
        "responses": {
          "200": {
            "description": "List of matching position titles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "Example success response",
                    "value": [
                      "Software Engineer",
                      "Senior Software Engineer",
                      "Software Engineering Intern"
                    ]
                  }
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example2": {
                    "summary": "Example error response",
                    "value": {
                      "message": "Database query failed"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "comp",
            "in": "query",
            "description": "Company name and optional position search term",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pos",
            "in": "query",
            "description": "Company name and optional position search term",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Search"
        ]
      }
    },
    "/autocomplete/location": {
      "get": {
        "deprecated": false,
        "summary": "Search for locations with fuzzy matching across city, state, and state code",
        "security": [],
        "responses": {
          "200": {
            "description": "List of matching locations in \"City, State_Code\" format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "Example success response",
                    "value": [
                      "Boston, MA",
                      "Cambridge, MA",
                      "Waltham, MA"
                    ]
                  }
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example2": {
                    "summary": "Example error response",
                    "value": {
                      "message": "Database query failed"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "loc",
            "in": "query",
            "description": "Search parameters for location",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Search"
        ]
      }
    },
...
  },
...
}